+++ metrics-core/src/main/java/com/yammer/metrics/core/MetricsRegistry.java
@@ -4,7 +4,6 @@ import com.yammer.metrics.core.Histogram.SampleType;
 import com.yammer.metrics.util.MetricPredicate;
 import com.yammer.metrics.util.ThreadPools;
 
-import javax.management.MalformedObjectNameException;
 import java.util.*;
 import java.util.concurrent.*;
 
@@ -64,68 +63,6 @@ public class MetricsRegistry {
     public <T> Gauge<T> newGauge(MetricName metricName,
                                        Gauge<T> metric) {
         return getOrAdd(metricName, metric);
-    }
-
-    /**
-     * Given a JMX MBean's object name and an attribute name, registers a gauge for that attribute
-     * under the given class and name.
-     *
-     * @param klass      the class which owns the metric
-     * @param name       the name of the metric
-     * @param objectName the object name of the MBean
-     * @param attribute  the name of the bean's attribute
-     * @return a new {@link JmxGauge}
-     * @throws MalformedObjectNameException if the object name is malformed
-     * @deprecated use {@link #newGauge(Class, String, Gauge)} and {@link JmxGauge} instead
-     */
-    @Deprecated
-    @SuppressWarnings({"UnusedDeclaration", "deprecation" })
-    public JmxGauge newJmxGauge(Class<?> klass,
-                                String name,
-                                String objectName,
-                                String attribute) throws MalformedObjectNameException {
-        return newJmxGauge(klass, name, null, objectName, attribute);
-    }
-
-    /**
-     * Given a JMX MBean's object name and an attribute name, registers a gauge for that attribute
-     * under the given class, name, and scope.
-     *
-     * @param klass      the class which owns the metric
-     * @param name       the name of the metric
-     * @param scope      the scope of the metric
-     * @param objectName the object name of the MBean
-     * @param attribute  the name of the bean's attribute
-     * @return a new {@link JmxGauge}
-     * @throws MalformedObjectNameException if the object name is malformed
-     * @deprecated use {@link #newGauge(Class, String, String, Gauge)} and {@link JmxGauge} instead
-     */
-    @Deprecated
-    @SuppressWarnings("deprecation")
-    public JmxGauge newJmxGauge(Class<?> klass,
-                                String name,
-                                String scope,
-                                String objectName,
-                                String attribute) throws MalformedObjectNameException {
-        return newJmxGauge(createName(klass, name, scope), objectName, attribute);
-    }
-
-    /**
-     * Given a JMX MBean's object name and an attribute name, registers a gauge for that attribute
-     * under the given metric name.
-     *
-     * @param metricName the name of the metric
-     * @param objectName the object name of the MBean
-     * @param attribute  the name of the bean's attribute
-     * @return a new {@link JmxGauge}
-     * @throws MalformedObjectNameException if the object name is malformed
-     * @deprecated use {@link #newGauge(MetricName, Gauge)} and {@link JmxGauge} instead
-     */
-    @Deprecated
-    public JmxGauge newJmxGauge(MetricName metricName,
-                                String objectName,
-                                String attribute) throws MalformedObjectNameException {
-        return getOrAdd(metricName, new JmxGauge(objectName, attribute));
     }
 
     /**