+++ CHANGELOG.md
@@ -44,6 +44,7 @@ v2.0.0-BETA19: TBD
 * Renamed `VirtualMachineMetrics.GarbageCollector` to `VirtualMachineMetrics.GarbageCollectorStats`.
 * Moved Guice/Servlet support from `metrics-servlet` to `metrics-guice`.
 * Removed `metrics-aop`.
+* Removed `newJmxGauge` from both `Metrics` and `MetricsRegistry`. Just use `JmxGauge`.
 
 
 v2.0.0-BETA18: Dec 16 2011